Added support for "--daemon" flag.
Daemonizing is optional. This flag should be used when the service
should be used as daemon in the background. By default the service
will be run in foreground. In case you use some process management
external to your service you dont need any stuff to do unless
a package is deployed on your linux distribution and you probably
want to use init.d script (which is not yet provided).